requirements.txt
sounddevice
soundfile
requests
pyttsx3
pandas
python-dotenv
speechrecognition